python3 script added!
This includes examples of polynomials whose roots are correctly retrieved by this algorithm. Hopefully those polynomials are not alone: this algorithm can retrieve the roots of many more polynomials (not included in the examples!) ...

However I am certainly not responsible if this algorithm does not retrieve the roots of the polynomial you are feeding it.

If I feel particularly generous and positive minded I may find the time to look whether this algorithm is buggy or not. But this is clearly not 100% sure , sorry :/

Hope this algorithm can be useful to someone else than me at some point though.

See you (around)